+++ tpl-home.php
@@ -2,12 +2,9 @@
 /* Template Name: Home Page */
 
 /**
- * File Name tpl-home.php
  * @package WordPress
- * @subpackage ParentTheme
+ * @subpackage ChildTheme
  * @license GPL v2 - http://www.gnu.org/licenses/old-licenses/gpl-2.0.html
- * @version 1.8
- * @updated 01.20.14
  **/
 #################################################################################################### */
 
@@ -19,7 +16,7 @@ get_template_part( 'header' );
 		<?php get_template_part( 'loop-default' ); ?>
 	</div>
 	<div class="span4">
-		<?php vc_sidebars( 'Primary Sidebar' ); ?>
+		<?php get__widget_area( 'Primary Sidebar' ); ?>
 	</div>
 </div>
 <?php